Trait pallet_dkg_proposal_handler::pallet::Config

pub trait Config: Config + CreateSignedTransaction<Call<Self>> + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type OffChainAuthId: AppCrypto<Self::Public, Self::Signature>;
    type SignedProposalHandler: OnSignedProposal<Self::MaxProposalLength>;
    type BatchId: Member + Parameter + Default + Encode + Decode + AtLeast32BitUnsigned + MaxEncodedLen + Copy;
    type MaxProposalsPerBatch: Get<u32> + Debug + Clone + Eq + PartialEq + PartialOrd + Ord + TypeInfo;
    type UnsignedProposalExpiry: Get<BlockNumberFor<Self>>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ValidatorSet: ValidatorSetWithIdentification<Self::AccountId>;
    type ReportOffences: ReportOffence<Self::AccountId, IdentificationTuple<Self>, DKGMisbehaviourOffence<IdentificationTuple<Self>>>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an RuntimeEvent.

type OffChainAuthId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.

type SignedProposalHandler: OnSignedProposal<Self::MaxProposalLength>

The signed proposal handler trait

type BatchId: Member + Parameter + Default + Encode + Decode + AtLeast32BitUnsigned + MaxEncodedLen + Copy

type MaxProposalsPerBatch: Get<u32> + Debug + Clone + Eq + PartialEq + PartialOrd + Ord + TypeInfo

Max number of signed proposal submissions per batch;

type UnsignedProposalExpiry: Get<BlockNumberFor<Self>>

Max blocks to store an unsigned proposal

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may forcibly reset parameters or otherwise alter privileged attributes.

type ValidatorSet: ValidatorSetWithIdentification<Self::AccountId>

A type for retrieving the validators supposed to be online in a session.

type ReportOffences: ReportOffence<Self::AccountId, IdentificationTuple<Self>, DKGMisbehaviourOffence<IdentificationTuple<Self>>>

A type that gives us the ability to submit offence reports for DKG misbehaviours

type WeightInfo: WeightInfo

Pallet weight information